requirements.txt
# SC Signature Scanner
# Star Citizen signature detection and identification tool

pytesseract>=0.3.10
Pillow>=10.0.0
watchdog>=3.0.0
pystray>=0.19.0
keyboard>=0.13.5
numpy>=1.24.0
requests>=2.31.0
websockets>=12.0
